requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.65.4
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
python-socks==2.7.1
sniffio==1.3.1
socksio==1.0.0
starlette==0.46.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
